requirements.txt
# Core bot
python-telegram-bot==20.*
nest_asyncio

# AI
anthropic
openai

# Google APIs
google-auth
google-auth-httplib2
google-api-python-client

# Document parsing
python-docx

# Utilities
python-dotenv
